+++ src/main/java/dev/eliezer/superticket/controller/ClientRestController.java
@@ -2,6 +2,7 @@ package dev.eliezer.superticket.controller;
 
 import dev.eliezer.superticket.domain.model.Client;
 import dev.eliezer.superticket.service.ClientService;
+import dev.eliezer.superticket.service.exception.BusinessException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -9,6 +10,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -51,7 +53,11 @@ public record ClientRestController(ClientService clientService) {
             @Content(schema = @Schema(implementation = Client.class))})
     @ApiResponse(responseCode = "422", description = "Invalid client data provided", content = {
             @Content(schema = @Schema(implementation = Object.class))})
-    public ResponseEntity<Client> insert(@Valid @RequestBody Client clientToInsert){
+    public ResponseEntity<Client> insert(@Valid @RequestBody Client clientToInsert, HttpServletRequest request){
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole != 2) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         var clientInserted = clientService.insert(clientToInsert);
         URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
@@ -68,7 +74,11 @@ public record ClientRestController(ClientService clientService) {
             @Content(schema = @Schema(implementation = Object.class))})
     @ApiResponse(responseCode = "422", description = "Invalid client data provided", content = {
             @Content(schema = @Schema(implementation = Object.class))})
-    public ResponseEntity<Client> update(@Valid @PathVariable Long id, @RequestBody Client clientToUpdate){
+    public ResponseEntity<Client> update(@Valid @PathVariable Long id, @RequestBody Client clientToUpdate, HttpServletRequest request){
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole != 2) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         var clientUpdated = clientService.update(id, clientToUpdate);
         URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
@@ -82,7 +92,11 @@ public record ClientRestController(ClientService clientService) {
     @ApiResponse(responseCode = "200", description = "Client successfully deleted")
     @ApiResponse(responseCode = "404", description = "Client not found")
 
-    public ResponseEntity<String> delete(@PathVariable Long id){
+    public ResponseEntity<String> delete(@PathVariable Long id, HttpServletRequest request){
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole != 2) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         clientService.delete(id);
         return ResponseEntity.ok("client successfully deleted");
     }
