+++ src/main/java/dev/eliezer/superticket/controller/StatusRestController.java
@@ -2,6 +2,7 @@ package dev.eliezer.superticket.controller;
 
 import dev.eliezer.superticket.domain.model.Status;
 import dev.eliezer.superticket.service.StatusService;
+import dev.eliezer.superticket.service.exception.BusinessException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -10,6 +11,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.MediaType;
@@ -52,13 +54,17 @@ public record StatusRestController(StatusService statusService) {
             @Content(schema = @Schema(implementation = Status.class))})
     @ApiResponse(responseCode = "422", description = "Invalid status data provided", content = {
             @Content(schema = @Schema(implementation = Object.class))})
-    public ResponseEntity<Status> insert(@Valid @RequestBody Status statusToInsert) {
-            var statusInserted = statusService.insert(statusToInsert);
-            URI location = ServletUriComponentsBuilder.fromCurrentRequest()
-                    .path("/{id}")
-                    .buildAndExpand(statusInserted.getId())
-                    .toUri();
-            return ResponseEntity.created(location).body(statusInserted);
+    public ResponseEntity<Status> insert(@Valid @RequestBody Status statusToInsert, HttpServletRequest request) {
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole != 2) {
+            throw new BusinessException("Unauthorized Access.");
+        }
+        var statusInserted = statusService.insert(statusToInsert);
+        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
+                 .path("/{id}")
+                 .buildAndExpand(statusInserted.getId())
+                 .toUri();
+        return ResponseEntity.created(location).body(statusInserted);
 
         }
 
@@ -70,7 +76,11 @@ public record StatusRestController(StatusService statusService) {
             @Content(schema = @Schema(implementation = Object.class))})
     @ApiResponse(responseCode = "422", description = "Invalid status data provided", content = {
             @Content(schema = @Schema(implementation = Object.class))})
-    public ResponseEntity<Status> update(@PathVariable Long id, @RequestBody Status statusToUpdate){
+    public ResponseEntity<Status> update(@PathVariable Long id, @RequestBody Status statusToUpdate, HttpServletRequest request){
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole != 2) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         var statusUpdated = statusService.update(id, statusToUpdate);
         URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
@@ -85,7 +95,11 @@ public record StatusRestController(StatusService statusService) {
             @ApiResponse(responseCode = "200", description = "Status successfully deleted"),
             @ApiResponse(responseCode = "404", description = "Status not found")
     })
-    public ResponseEntity<String> delete(@PathVariable Long id){
+    public ResponseEntity<String> delete(@PathVariable Long id, HttpServletRequest request){
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole != 2) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         statusService.delete(id);
         return ResponseEntity.ok("status successfully deleted");
     }
