+++ src/main/java/dev/eliezer/superticket/controller/TicketAnnotationRestController.java
@@ -4,6 +4,7 @@ import dev.eliezer.superticket.domain.model.TicketAnnotation;
 import dev.eliezer.superticket.dto.TicketAnnotationRequestDTO;
 import dev.eliezer.superticket.dto.TicketAnnotationResponseDTO;
 import dev.eliezer.superticket.service.TicketAnnotationService;
+import dev.eliezer.superticket.service.exception.BusinessException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -12,6 +13,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -52,7 +54,11 @@ public record TicketAnnotationRestController(TicketAnnotationService ticketAnnot
             @Content(schema = @Schema(implementation = TicketAnnotation.class))})
     @ApiResponse(responseCode = "422", description = "Invalid annotation data provided", content = {
             @Content(schema = @Schema(implementation = Object.class))})
-    public ResponseEntity<TicketAnnotationResponseDTO> insert(@Valid @RequestBody TicketAnnotationRequestDTO request){
+    public ResponseEntity<TicketAnnotationResponseDTO> insert(@Valid @RequestBody TicketAnnotationRequestDTO request, HttpServletRequest httpRequest){
+        Long userRole = Long.valueOf(httpRequest.getAttribute("user_role").toString());
+        if (userRole == 3) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         var annotationInserted = ticketAnnotationService.insert(request);
         URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
@@ -68,7 +74,11 @@ public record TicketAnnotationRestController(TicketAnnotationService ticketAnnot
             @ApiResponse(responseCode = "200", description = "Ticket annotation successfully deleted"),
             @ApiResponse(responseCode = "404", description = "Ticket annotation not found")
     })
-    public ResponseEntity<String> delete(@PathVariable Long id){
+    public ResponseEntity<String> delete(@PathVariable Long id, HttpServletRequest request){
+        Long userRole = Long.valueOf(request.getAttribute("user_role").toString());
+        if (userRole == 3) {
+            throw new BusinessException("Unauthorized Access.");
+        }
         ticketAnnotationService.delete(id);
         return ResponseEntity.ok("ticket annotation successfully deleted");
     }
